requirements.txt
flask
python-dotenv
mysql-connector-python
werkzeug
pandas
openpyxl
